Sleeping mechanic - storyline - UI
Now, shelter can be used a single time in order to prevent the sleeping mechanic being overpowered. Furthermore, it can only be used during nights now. There is a storyline added to mysterious strange that tells the basics and purpose of the game. There are also changes about UI (messages etc.)